ベクトルを int8 に量子化して索引を 1.64GB から 508MB にする
候補生成の内積と Top-K を Rust に移し (`rust/src/gemv.rs`)、索引の
ベクトルを int8 で持つ (FORMAT_VERSION 4)。CSR の境界も int32 にした。

**目的はサイズであって速度ではない。** 内積は BLAS の float32 と互角で
(17〜29ms 対 15〜26ms)、読むバイト数が 1/4 になっても DRAM 帯域律速なので
速くならない。速度で得しているのは Top-K だけ (argpartition の 9〜16ms に
対し 3〜10ms)。

最初は「候補生成が 3.9 倍速い」と読んだが測定ミスだった。f32 (582MB) と
int8 (145MB) を同じプロセスで交互に測ると、f32 が毎回 L3 を流すので次に
走る int8 だけが有利に出る。プロセスを分けると両者は並ぶ。共有資源を
奪い合う 2 案の比較は必ず別プロセスで測る。

量子化の誤差は内積で最大 0.012 あり候補生成の順位を揺らすので、
DEFAULT_CANDIDATES を 5000 から 8000 に広げて吸収する (float32 基準の
再現率 0.919 -> 0.997)。受入テスト 47 件が実辞書の索引で通ることを確認した。

あわせて Top-K が同点を返すときの非決定性を直した。スコアだけで比べると
どの候補が k 件に残るかがスレッドの合流順で変わり、代表選びが揺れて同じ
クエリが違う表記を返す (全件同点の 5000 行から上位 10 件を取ると行 0, 1 を
飛ばして 2852 が入っていた)。行番号まで見て全順序にする。

+++ src/jpr/distance.py
@@ -478,7 +478,7 @@ def edit_distance_csr(
         np.ascontiguousarray(query_ids, dtype=np.int32),
         np.ascontiguousarray(rows, dtype=np.int64),
         np.ascontiguousarray(phoneme_ids_blob, dtype=np.uint8),
-        np.ascontiguousarray(phoneme_bounds, dtype=np.int64),
+        np.ascontiguousarray(phoneme_bounds, dtype=np.int32),
         np.ascontiguousarray(distance_ids, dtype=np.int32),
         _SUB_FLAT32,
         _INDEL_FLAT32,
